+++ lib/src/theme/theme.dart
@@ -1,1 +1,5 @@
+export 'draggable_panel_button_theme_data.dart';
+export 'draggable_panel_handle_theme_data.dart';
+export 'draggable_panel_item_theme_data.dart';
 export 'draggable_panel_theme.dart';
+export 'draggable_panel_tooltip_theme_data.dart';